requirements-jetson.txt
# Jetson / 边缘设备：请先用官方渠道安装 torch、torchvision（torchaudio 需与厂商 torch 同批次 wheel，否则勿装 PyPI 版）（见 JETSON_PYTORCH.md），
# 再安装本文件其余依赖，避免 pip 覆盖厂商 CUDA 版 PyTorch。
#
#   pip install -r requirements-jetson.txt

# 核心依赖
numpy==2.2.6
opencv-python==4.11.0.86
PyYAML==6.0.1
psutil==7.2.2

# 模型依赖（不含 torch 三件套）
ultralytics==8.4.21
funasr==1.3.1
addict>=2.4.0
ollama>=0.1.0
# Video-Depth-Anything（perception/vda）推理依赖
easydict>=1.10
einops>=0.4.1

# 其他依赖
Pillow==12.1.1
requests==2.32.5
tqdm==4.67.3
librosa==0.11.0
soundfile==0.13.1
pydub==0.25.1
protobuf==6.33.5
sympy==1.14.0

# 其他必要依赖
transformers==4.57.6
huggingface_hub==0.36.2
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
pandas==2.3.3
matplotlib==3.10.8

# 网络和API相关
fastapi==0.134.0
uvicorn==0.41.0
httpx==0.28.1

# 工具库
loguru==0.7.3
python-dotenv==1.2.2
